backend/requirements.txt
flask
flask-cors

langchain-core
langgraph
langgraph-checkpoint

langchain-huggingface
pydantic

python-dotenv
